+++ src/main/java/org/crops/fitserver/global/exception/FitException.java
@@ -1,5 +1,7 @@
 package org.crops.fitserver.global.exception;
 
+import org.crops.fitserver.global.http.ErrorType;
+
 public class FitException extends RuntimeException {
 
 	private ErrorType errorType;
