+++ src/main/java/org/crops/fitserver/global/exception/NotFoundException.java
@@ -1,5 +1,7 @@
 package org.crops.fitserver.global.exception;
 
+import org.crops.fitserver.global.http.ErrorType;
+
 public class NotFoundException extends FitException {
 
 	public NotFoundException() {
