+++ src/main/java/org/crops/fitserver/global/exception/UnauthorizedException.java
@@ -1,5 +1,7 @@
 package org.crops.fitserver.global.exception;
 
+import org.crops.fitserver.global.http.ErrorType;
+
 public class UnauthorizedException extends FitException {
 
 	public UnauthorizedException() {
