+++ src/main/java/org/crops/fitserver/global/feign/oauth/OAuthToken.java
@@ -1,17 +1,18 @@
-package org.crops.fitserver.global.feign;
+package org.crops.fitserver.global.feign.oauth;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class GoogleSocialUserProfile {
+public class OAuthToken {
 
-	private String sub;
-	private String email;
-	private String name;
+	private String accessToken;
 }
