+++ src/main/java/org/crops/fitserver/global/feign/oauth/google/GoogleSocialUserProfile.java
@@ -1,4 +1,4 @@
-package org.crops.fitserver.global.feign;
+package org.crops.fitserver.global.feign.oauth.google;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,7 +9,9 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class KakaoSocialUserProfile {
+public class GoogleSocialUserProfile {
 
-	private Long id;
+	private String sub;
+	private String email;
+	private String name;
 }
