+++ src/main/java/org/crops/fitserver/global/feign/oauth/kakao/KakaoSocialUserProfile.java
@@ -1,18 +1,15 @@
-package org.crops.fitserver.global.feign;
+package org.crops.fitserver.global.feign.oauth.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class OAuthToken {
+public class KakaoSocialUserProfile {
 
-	private String accessToken;
+	private Long id;
 }
