+++ src/main/java/org/crops/fitserver/global/jwt/JwtResolver.java
@@ -13,7 +13,7 @@ import java.util.Date;
 import javax.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.crops.fitserver.global.exception.ErrorType;
+import org.crops.fitserver.global.http.ErrorType;
 import org.crops.fitserver.global.exception.FitException;
 import org.crops.fitserver.global.exception.UnauthorizedException;
 import org.springframework.stereotype.Service;
